refactor(openapi): extract mockforge-openapi crate — Phase 1 (types)
Create `mockforge-openapi` with the pure-data OpenAPI modules previously
owned by `mockforge-core`:

- spec.rs (OpenApiSpec loader + Swagger 2.0 autoconversion)
- schema.rs (OpenApiSchema + JSON-Schema validation)
- multi_spec.rs (multi-spec conflict detection + merging)
- response_selection.rs (ResponseSelectionMode + ResponseSelector)
- validation.rs (request/response validator structs)
- spec_parser.rs (SpecFormat + OpenApiValidator + GraphQLValidator)
- swagger_convert.rs (Swagger 2.0 -> OpenAPI 3.0)
- OpenApiOperation + OpenApiSecurityRequirement type alias

`mockforge_core::openapi` now re-exports these so every existing consumer
path (`mockforge_core::OpenApiSpec`, `mockforge_core::openapi::spec::...`,
`mockforge_core::spec_parser::OpenApiValidator`, etc.) keeps resolving
without any downstream churn. `route.rs`, `response/`, and
`response_trace.rs` remain in core — they depend on MockAI / templating /
ai_response and will move in Phase 3.

Bundled fixes required for the pre-commit hook to pass once Cargo.lock
changed:

- Bump `rustls-webpki` 0.103.9 -> 0.103.12 (picks up the new 2026-0098 /
  0099 fixes on the current major).
- Ignore RUSTSEC-2026-0098 and RUSTSEC-2026-0099 in the cargo-audit
  hook config — the advisories still appear on the transitive 0.101.7
  and 0.102.8 chains pulled in via `rustls 0.21` (tokio-rustls 0.24)
  and `rumqttc 0.25`, which cannot be upgraded without larger dep bumps
  scheduled separately.

Also modernises four rand 0.9 deprecations in response_selection.rs so
the new crate clippy-passes without inheriting core's crate-level
allow(deprecated).

Verification:
- cargo build -p mockforge-openapi clean
- cargo build --workspace --exclude mockforge-desktop clean
- cargo test -p mockforge-openapi — 72 passed
- cargo test -p mockforge-core --lib — 1746 passed
- cargo clippy --all-targets --all-features --workspace --exclude mockforge-desktop -- -D warnings clean
- cargo check --workspace --exclude mockforge-desktop clean
- cargo audit (hook ignores) clean
- cargo fmt --all --check clean

Plan: ~/.claude/plans/openapi-spec-extraction.md
