+++ .pre-commit-config.yaml
@@ -72,7 +72,7 @@ repos:
   - id: cargo-audit
     name: cargo audit
     entry: cargo
-    args: [audit, --ignore, RUSTSEC-2023-0071, --ignore, RUSTSEC-2026-0020, --ignore, RUSTSEC-2026-0021, --ignore, RUSTSEC-2026-0037, --ignore, RUSTSEC-2026-0085, --ignore, RUSTSEC-2026-0086, --ignore, RUSTSEC-2026-0087, --ignore, RUSTSEC-2026-0088, --ignore, RUSTSEC-2026-0089, --ignore, RUSTSEC-2026-0091, --ignore, RUSTSEC-2026-0092, --ignore, RUSTSEC-2026-0094, --ignore, RUSTSEC-2026-0095, --ignore, RUSTSEC-2026-0096, --ignore, RUSTSEC-2026-0002, --ignore, RUSTSEC-2026-0093]
+    args: [audit, --ignore, RUSTSEC-2023-0071, --ignore, RUSTSEC-2026-0020, --ignore, RUSTSEC-2026-0021, --ignore, RUSTSEC-2026-0037, --ignore, RUSTSEC-2026-0085, --ignore, RUSTSEC-2026-0086, --ignore, RUSTSEC-2026-0087, --ignore, RUSTSEC-2026-0088, --ignore, RUSTSEC-2026-0089, --ignore, RUSTSEC-2026-0091, --ignore, RUSTSEC-2026-0092, --ignore, RUSTSEC-2026-0094, --ignore, RUSTSEC-2026-0095, --ignore, RUSTSEC-2026-0096, --ignore, RUSTSEC-2026-0002, --ignore, RUSTSEC-2026-0093, --ignore, RUSTSEC-2026-0098, --ignore, RUSTSEC-2026-0099]
     language: system
     pass_filenames: false
     files: ^Cargo\.(toml|lock)$
